@@ -6,25 +6,24 @@
     "authors": [
         {
             "name": "Devon Weller",
-            "email": "dweller@devonweller.com",
-            "homepage": "http://devonweller.com"
+            "email": "devon@tokenly.com",
+            "homepage": "https://tokenly.com"
         }
     ],
     "require": {
-        "php": ">=5.5.0",
-        "tokenly/xcpd-client": "dev-master"
+        "php": ">=5.6.0",
+        "tokenly/xcpd-client": "^1"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4",
+        "phpunit/phpunit": "~6",
         "mockery/mockery": "~0.9",
-        "orchestra/testbench": "3.0.*@dev",
-        "laravel/framework": "5.0.x-dev@dev"
+        "orchestra/testbench": "^3.5"
     },
     "minimum-stability": "dev",
     "autoload": {
         "classmap": [
             "test/base/TestCase.php"
         ],
-        "psr-0": {"Tokenly\\CounterpartyAssetInfoCache": "src"}
+        "psr-4": {"Tokenly\\CounterpartyAssetInfoCache\\": "src/"}
     }
 }
